+++ backend/src/main/java/org/tarjetamish/transaction/mapper/ITransactionConverter.java
@@ -1,0 +1,10 @@
+package org.tarjetamish.transaction.mapper;
+
+import org.tarjetamish.transaction.dto.TransactionDTO;
+import org.tarjetamish.transaction.model.Transaction;
+
+public interface ITransactionConverter {
+
+    TransactionDTO toTransactionDTO(Transaction transaction);
+    Transaction toTransaction(TransactionDTO transactionDTO);
+}
